base: Add Ambient Customizations [1/2]
This adds the option to add custom text and a custom image on the ambient screen. This option is still in development, so be patient, the bugs that arise will be resolved.
Based on OmniROM Edge Lighting and Custom Headers.

SDB: adapt for R & put it behind notification lights

DerpFest: Use our fonts and catch up with SynthOS

Good Work by TikkiX2

Change-Id: I8bbda7e908ef4bae892c59fdbcf6032a799b3f5b